+++ src/components/manage/Contents/Contents.jsx
@@ -50,7 +50,6 @@ import {
   moveContent,
   orderContent,
   sortContent,
-  addMessage,
 } from '../../../actions';
 import { getBaseUrl } from '../../../helpers';
 import Indexes from '../../../constants/Indexes';
@@ -64,8 +63,10 @@ import {
   ContentsPropertiesModal,
   Pagination,
   Toolbar,
+  Toast,
   Icon as IconNext,
 } from '../../../components';
+import { toast } from 'react-toastify';
 
 import backSVG from '../../../icons/back.svg';
 
@@ -155,6 +156,10 @@ const messages = defineMessages({
     id: 'Upload',
     defaultMessage: 'Upload',
   },
+  success: {
+    id: 'Success',
+    defaultMessage: 'Success',
+  },
 });
 
 /**
@@ -179,7 +184,6 @@ class Contents extends Component {
     moveContent: PropTypes.func.isRequired,
     orderContent: PropTypes.func.isRequired,
     sortContent: PropTypes.func.isRequired,
-    addMessage: PropTypes.func.isRequired,
     clipboardRequest: PropTypes.shape({
       loading: PropTypes.bool,
       loaded: PropTypes.bool,
@@ -735,10 +739,12 @@ class Contents extends Component {
   cut(event, { value }) {
     this.props.cut(value ? [value] : this.state.selected);
     this.onSelectNone();
-    this.props.addMessage(
-      null,
-      this.props.intl.formatMessage(messages.messageCut),
-      'success',
+    toast.success(
+      <Toast
+        success
+        title={this.props.intl.formatMessage(messages.success)}
+        content={this.props.intl.formatMessage(messages.messageCut)}
+      />,
     );
   }
 
@@ -752,10 +758,12 @@ class Contents extends Component {
   copy(event, { value }) {
     this.props.copy(value ? [value] : this.state.selected);
     this.onSelectNone();
-    this.props.addMessage(
-      null,
-      this.props.intl.formatMessage(messages.messageCopied),
-      'success',
+    toast.success(
+      <Toast
+        success
+        title={this.props.intl.formatMessage(messages.success)}
+        content={this.props.intl.formatMessage(messages.messageCopied)}
+      />,
     );
   }
 
@@ -846,10 +854,12 @@ class Contents extends Component {
         getBaseUrl(this.props.pathname),
       );
     }
-    this.props.addMessage(
-      null,
-      this.props.intl.formatMessage(messages.messagePasted),
-      'success',
+    toast.success(
+      <Toast
+        success
+        title={this.props.intl.formatMessage(messages.success)}
+        content={this.props.intl.formatMessage(messages.messagePasted)}
+      />,
     );
   }
 
@@ -1304,7 +1314,6 @@ export default compose(
       moveContent,
       orderContent,
       sortContent,
-      addMessage,
     },
   ),
 )(Contents);