+++ src/components/theme/ContactForm/ContactForm.jsx
@@ -12,8 +12,10 @@ import { Portal } from 'react-portal';
 import { Container, Message, Icon } from 'semantic-ui-react';
 import { defineMessages, injectIntl, intlShape } from 'react-intl';
 import { Link, withRouter } from 'react-router-dom';
-import { Form, Toolbar } from '../..';
-import { addMessage, emailNotification } from '../../../actions';
+import { toast } from 'react-toastify';
+
+import { Form, Toolbar, Toast } from '../../';
+import { emailNotification } from '../../../actions';
 import { getBaseUrl } from '../../../helpers';
 
 const messages = defineMessages({
@@ -57,6 +59,10 @@ const messages = defineMessages({
     id: 'Back',
     defaultMessage: 'Back',
   },
+  success: {
+    id: 'Success',
+    defaultMessage: 'Success',
+  },
 });
 
 /**
@@ -72,7 +78,6 @@ class ContactForm extends Component {
    */
   static propTypes = {
     emailNotification: PropTypes.func.isRequired,
-    addMessage: PropTypes.func.isRequired,
     intl: intlShape.isRequired,
     error: PropTypes.shape({
       message: PropTypes.string,
@@ -113,10 +118,12 @@ class ContactForm extends Component {
    */
   componentWillReceiveProps(nextProps) {
     if (this.props.loading && nextProps.loaded) {
-      this.props.addMessage(
-        null,
-        this.props.intl.formatMessage(messages.messageSent),
-        'success',
+      toast.success(
+        <Toast
+          success
+          title={this.props.intl.formatMessage(messages.success)}
+          content={this.props.intl.formatMessage(messages.saved)}
+        />,
       );
     }
   }
@@ -235,6 +242,6 @@ export default compose(
       error: state.emailNotification.error,
       pathname: props.location.pathname,
     }),
-    { emailNotification, addMessage },
+    { emailNotification },
   ),
 )(ContactForm);